Reorganize namespaces and directories
