+++ src/AudioTagger.Console/GetUserInput.cs
@@ -1,6 +1,7 @@
 ﻿using System;
 using System.Linq;
 using System.Collections.Generic;
+using AudioTagger.Library;
 
 namespace AudioTagger.Console;
