+++ src/AudioTagger.Console/OperationLibrary.cs
@@ -1,5 +1,7 @@
-using System.Diagnostics.CodeAnalysis;
+using AudioTagger.Console.Operations;
 using OperationFlags = System.Collections.Generic.HashSet<string>;
+using System.Diagnostics.CodeAnalysis;
+using AudioTagger.Library;
 
 namespace AudioTagger.Console;
