+++ src/AudioTagger.Console/Operations/MediaFileRenamer.cs
@@ -3,7 +3,7 @@ using System.Text.RegularExpressions;
 using Spectre.Console;
 using AudioTagger.Library;
 
-namespace AudioTagger.Console;
+namespace AudioTagger.Console.Operations;
 
 public sealed class MediaFileRenamer : IPathOperation
 {
