+++ src/AudioTagger.Console/Operations/TagScanner.cs
@@ -1,7 +1,8 @@
 using System.Text;
 using System.Text.RegularExpressions;
+using AudioTagger.Library;
 
-namespace AudioTagger.Console;
+namespace AudioTagger.Console.Operations;
 
 public sealed class TagScanner : IPathOperation
 {
