+++ src/AudioTagger.Console/Operations/TagStats.cs
@@ -1,8 +1,9 @@
 using System.Diagnostics.CodeAnalysis;
 using System.Text.RegularExpressions;
+using AudioTagger.Library;
 using Spectre.Console;
 
-namespace AudioTagger.Console;
+namespace AudioTagger.Console.Operations;
 
 public sealed class TagStats : IPathOperation
 {
