+++ src/AudioTagger.Tests/OutputLineTests.cs
@@ -1,4 +1,5 @@
-﻿using Xunit;
+﻿using AudioTagger.Library;
+using Xunit;
 
 namespace AudioTagger.Tests;
